App for creating and visualising a time series
Allows to select dates in the calendar and input a value for the date.
Line chart shows all values and calculate and average with a moving
window. Window size can be set between 3-11 with a step size of 2.

App uses server side rendereing and stores the data on the server.